add some casts to avoid compile warnings

+++ tinyexpr.c
@@ -237,8 +237,8 @@ void next_token(state *s) {
         start = s->next;
         while ((s->next[0] >= 'a' && s->next[0] <= 'z') || (s->next[0] >= '0' && s->next[0] <= '9') || (s->next[0] == '_')) s->next++;
         
-        const te_variable *var = find_lookup(s, start, s->next - start);
-        if (!var) var = find_builtin(start, s->next - start);
+        const te_variable *var = find_lookup(s, start, (int)(s->next - start));
+        if (!var) var = find_builtin(start, (int)(s->next - start));
         
         if (!var) {
           s->type = TOK_ERROR;
@@ -579,7 +579,7 @@ te_expr *te_compile(const char *expression, const te_variable *variables, int va
   if (s.type != TOK_END) {
     te_free(root);
     if (error) {
-      *error = (s.next - s.start);
+      *error = (int)(s.next - s.start);
       if (*error == 0) *error = 1;
     }
     return 0;
